tasks for bogabot
