+++ sandbox/boletin.py
@@ -2,206 +2,13 @@
 from bs4 import BeautifulSoup
 from pathlib import Path
 import json
-import gzip
 import requests
 from requests.adapters import HTTPAdapter, Retry
 import markdown
 import sys
 
-
-def load_json_gz(fpath):
-    with gzip.open(fpath,'rt',encoding='utf-8') as jf:
-        return json.load(jf)
-
-MODEL_DUMB_BRAIN="huihui_ai/qwen2.5-1m-abliterated:14b"
-MODEL_SMART_BRAIN="qwq:32b"
-
-law_ref = load_json_gz(Path('../data/leyes_ref.json.gz'))
-decree_ref = load_json_gz(Path('../data/decretos_ref.json.gz'))
-with open("../data/mapa_context.txt","r",encoding="utf-8") as fp:
-    mapa_context = f"Estos son los cargos conocidos:\n'''{fp.read()}\n'''\n\n"
-
-
-prompt_token_total = 0
-eval_token_total = 0
-gpu_time_total = 0
-
-ollama_session = requests.Session()
-#ollama_session.get("https://italia-bibliography-objectives-preference.trycloudflare.com/?token=d12913169a9666262987b7792bbbc08f251e331c571721e8f67b1c3jaja_de_aca")
-
-def query_ollama(model_name, context, query, max_context=512*1024):
-    try:
-        global prompt_token_total
-        global eval_token_total
-        global gpu_time_total
-        # Construct the Ollama API URL
-        base_url = "http://localhost:11434/api/generate"
-        base_url = "https://italia-bibliography-objectives-preference.trycloudflare.com/api/generate"
-
-        # Prepare the prompt by combining context and query
-        prompt = f"Contexto:\n'''\n{context}\n'''\n\nTarea:\n'''\n{query}\n'''\n"
-
-        # Prepare the payload for the request
-        payload = {
-            "model": model_name,
-            "prompt": prompt,
-            "stream": False,
-            "options": {
-                # Bogabot personality 1.0
-                "temperature": 0.4,
-                "seed": 42, # guaranteed to be random
-                "top_k": 23,
-                "top_p": 0.5,
-                "num_ctx": 8192, # Testing
-            }
-        }
-
-        # Make the POST request to Ollama API
-        response = ollama_session.post(base_url, json=payload, verify=False)
-        response.raise_for_status()
-
-        # Extract and return the generated text from the response
-        data = response.json()
-        prompt_token_total += data['prompt_eval_count']
-        eval_token_total += data['eval_count']
-        gpu_time_total += data['total_duration']
-        llm_response = data["response"]
-        is_thinking = llm_response.find("<think>")
-        end_of_think = llm_response.find("</think>")
-        if end_of_think > 0:
-            llm_response = llm_response[(end_of_think+8):]
-        if is_thinking >= 0 and end_of_think < 0:
-            print("--------------------------------------------------")
-            print("--------------------------------------------------")
-            print("--------------------------------------------------")
-            print(f"Input:\n{prompt}\n")
-            print("--------------------------------------------------")
-            print("--------------------------------------------------")
-            print("--------------------------------------------------")
-            
-        return llm_response
-
-    except requests.exceptions.RequestException as e:
-        print(f"An error occurred while communicating with Ollama API: {e}")
-        return None
-
-
-def get_details(url, session, do_brief=True, do_ref=True):
-    print(f"fetching details from {url}")
-    with session.get(url) as r:
-        html_content = r.text
-
-
-    soup = BeautifulSoup(html_content, 'html.parser')
-
-    for script in soup(['script', 'meta', 'link', 'style']):
-        script.decompose()  # Removes the tag completely from the tree
-
-    cuerpo_div = soup.find('div', {'id': 'cuerpoDetalleAviso'}).contents[1]
-    body = str(cuerpo_div) if cuerpo_div else None  # Preserve the HTML structure
-    
-
-    data = {
-        "full_text": body
-    }
-    
-    if do_brief:
-        response = query_ollama(MODEL_SMART_BRAIN, mapa_context + body, "Crea un resumen del siguiente texto, siempre menciona a todos los firmantes que son los nombres al final, si es una designacion solo menciona a las personas involucradas y sus roles, si hay datos tabulados solo menciona su existencia, no ofrezcas mas ayuda, la respuesta es final, el resumen no debe tener mas de 600 caracteres, ignorar tags HTML.")
-        #refined_response = query_ollama(MODEL, response, "Corrije errores ortograficos en el siguiente texto. Si no hay errores solo copia el texto. No ofrezcas mas ayuda ni hagas aclaraciones")
-        data['brief'] = response
-        print(f"BRIEF:\n\n{response}\n\n.-.-.\n")
-    
-    if do_ref:
-        raw_law_list = query_ollama(MODEL_DUMB_BRAIN, body, "Crear una lista en formato JSON de numeros de ley (sin articulos). Limitaciones: - Solo deben ser leyes, ignorar decretos, resoluciones, comunicaciones u otro tipo de normas. - Sin comentarios. - Sin repetidos - Sin detalles - En caso de no existir leyes mencionadas la respuesta es un vector vacio: '[]'. - No incluir markdown para indicar que es JSON.")
-        print(raw_law_list)
-        try:
-            law_list = json.loads(raw_law_list)
-        except json.decoder.JSONDecodeError:
-            law_list = []
-            print(f"JSON error in law_list !!!!!!!!! {raw_law_list}")
-
-        raw_decree_list = query_ollama(MODEL_DUMB_BRAIN, body, "Crear una lista en JSON de decretos mencionados, el formato es '\"123/2024\"' donde '123' es el numero de decreto y '2024' el año. Reglas: - No incluir leyes, resoluciones ni otro tipo de normas. - Sin comentarios. - Sin repetidos - Sin detalles. - Si no se mencionan decretos la respuesta es un vector vacio: '[]'. - No incluir markdown para indicar que es JSON. No pensarlo demasiado.")
-        print(raw_decree_list)
-        try:
-            decree_list = json.loads(raw_decree_list)
-        except json.decoder.JSONDecodeError:
-            decree_list = []
-            print(f"JSON error in decree_list !!!!!!!!!: {raw_decree_list}")
-
-        data['ref'] = "<ul>\n"
-        info_legs = set()  # lol
-        if len(law_list) > 0:
-            data['ref'] += "<li>Leyes:<ul>\n"
-            for law in law_list:
-                law_str = str(law).replace('.','')
-                if law_str.find('/') > 0:
-                    continue  # Dumb AI puts decree as law
-                data['ref'] += f"<li>{law_str}"
-                if law_str in law_ref:
-                    data['ref']+= "<ul>\n"
-                    matches = law_ref[law_str]
-                    for year in matches:
-                        for law_data in matches[year]:
-                            data['ref'] += f"<li>infoleg {law_str} - {year} - <a href=https://servicios.infoleg.gob.ar/infolegInternet/verNorma.do?id={law_data['id']}>{law_data['id']}</a>: {law_data['resumen']}</li>\n"
-                            info_legs.add(str(law_data['id']))
-                    data['ref'] += "</ul>"
-                data['ref'] += "</li>\n"
-            data['ref'] += "</ul></li>\n"
-        if len(decree_list) > 0:
-            data['ref'] += "<li>Decretos:<ul>\n"
-            for dec in decree_list:
-                dec_str = str(dec)
-                data['ref'] += f"<li>{dec_str}"
-                if '/' in dec_str:
-                    dec_n = dec_str.split('/')[0]
-                    dec_y = dec_str.split('/')[1]
-                    if dec_n not in decree_ref:
-                        if len(dec_n) == 4 and dec_n.startswith('20'):
-                            dec_n = dec_n[2:]
-                    if dec_n in decree_ref:
-                        decree_n = decree_ref[dec_n]
-                        if dec_y in decree_n:
-                            data['ref']+= "<ul>\n"
-                            matches = decree_n[str(dec_y)]
-                            for decree in matches:
-                                data['ref'] += f"<li>infoleg {dec_str} - <a href=https://servicios.infoleg.gob.ar/infolegInternet/verNorma.do?id={decree['id']}>{decree['id']}</a>: {decree['resumen']}</li>\n"
-                                info_legs.add(str(decree['id']))
-                            data['ref'] += "</ul>"
-                        elif dec_y in ['25','2025']:
-                            data['ref'] += " _TODO_preload_2025_dec_"
-                data['ref'] += "</li>\n"
-            data['ref'] += "</ul></li>\n"
-        data['ref'] += "</ul>\n"
-
-        print(data['ref'])
-        # now let's make a killer query with all references :D
-        full_context = []
-        for infoleg_id in info_legs:
-            infoleg_file = Path(f"../data/infoleg_html/{infoleg_id[-1]}/{infoleg_id}.html")
-            if infoleg_file.exists():
-                with open(infoleg_file, 'r', encoding='utf-8') as infoleg_html:
-                    info_soup = BeautifulSoup(infoleg_html, 'html.parser')
-                    for div in info_soup.find_all('div'):
-                        text = div.text
-                        if len(text.strip())>1:
-                            full_context.append(text)
-                    print(f"context loaded: {infoleg_file}")
-            else:
-                print(f"context not found: {infoleg_file}")
-        context_as_text = ""
-        for context in full_context:
-            context_as_text += context
-            context_as_text += "\n\n-------\n\n"
-        context_as_text += mapa_context
-        context_as_text += "Norma actual:\n"
-        context_as_text += body
-        print("-------------------------------------------------------")
-        print(f"Tamaño del contexto a enviar: {len(context_as_text)} bytes")
-        print("-------------------------------------------------------")
-        analysis = query_ollama(MODEL_SMART_BRAIN, context_as_text, "Explicar como la norma actual (la ultima en la lista) afecta o impacta sobre las normas anteriores. En caso de modificar leyes anteriores explicar los beneficios afectados de la ley anterior. En caso de no tener impacto o afectarlas no hace falta explicar.")
-        data['analysis'] = analysis
-        print(analysis)
-    return data
+tasks_path = Path('../tasks/')
+tasks_path.mkdir(exist_ok=True)
 
 
 def get_day(day:int ,month:int , year:int):
@@ -221,7 +28,7 @@ def get_day(day:int ,month:int , year:int):
     soup = BeautifulSoup(html_content, 'html.parser')
 
     results = {}
-
+    order = 0
     # Find all div elements with class "col-md-12"
     for item_div in soup.find_all('div', {'class': 'col-md-12'}):
         # Extract the link from the anchor tag
@@ -258,16 +65,18 @@ def get_day(day:int ,month:int , year:int):
         element = {
             'subject': subject,
             'link': link,
+            'order': order,
             'name': name,
             'official_id': official_id,
-            'data': get_details(f"https://www.boletinoficial.gob.ar{link}", session, do_ref=do_ref)
+            'data_link': f"https://www.boletinoficial.gob.ar{link}",
         }
+        task_path = tasks_path / f"bo_{year}_{month:02}_{day:02}_el_{link.split('/')[-2]}.json"
+        if task_path.exists() == False:
+            with open(task_path, 'w', encoding='utf-8') as task_file:
+                json.dump(element, task_file, indent=2, ensure_ascii=False)
+            
         results[link] = element
         print(f"\n\n-- subject: {subject} name: {name} id: {official_id}")
-    global prompt_token_total
-    global eval_token_total
-    global gpu_time_total
-    print(f"\n----------------------------------------\nTotal gpu time: {gpu_time_total} total prompt tokens: {prompt_token_total} total eval tokens: {eval_token_total}\n")
     return results
 
 
@@ -275,16 +84,10 @@ def get_day(day:int ,month:int , year:int):
 day= int(sys.argv[1])
 month= int(sys.argv[2])
 year= int(sys.argv[3])
-json_file_path = Path(f"bo{year-2000}{month:02}{day:02}.json")
 
-if json_file_path.exists() == False:
-    results = get_day(day,month,year)
-    with open(json_file_path, "w", encoding='utf-8') as json_file:
-        json.dump(results, json_file, indent=2, ensure_ascii=False)
-else: # Loads existing data if available
-    with open(json_file_path, 'r', encoding='utf-8') as json_file:
-        results = json.load(json_file)
+get_day(day,month,year)
 
+sys.exit(0)
 # Gen HTML
 
 with open(f"bo{year-2000}{month:02}{day:02}.html",'w') as html_o:
